basic/python/django-rest/requirements.txt
django
djangorestframework
markdown
django-filter
debugpy